+++ includes/parsers/ImageOverlayParser.php
@@ -2,12 +2,10 @@
 
 namespace Maps;
 
-use DataValues\Geo\Parsers\GeoCoordinateParser;
-use DataValues\Geo\Parsers\LatLongParser;
 use Maps\Elements\ImageOverlay;
 use Maps\Elements\WmsOverlay;
 use ValueParsers\ParseException;
-use ValueParsers\StringValueParser;
+use ValueParsers\ValueParser;
 
 /**
  * @since 3.1
@@ -15,17 +13,23 @@ use ValueParsers\StringValueParser;
  * @licence GNU GPL v2+
  * @author Jeroen De Dauw < jeroendedauw@gmail.com >
  */
-class ImageOverlayParser extends StringValueParser {
+class ImageOverlayParser implements ValueParser {
+
+	private $geocoder;
+
+	public function __construct() {
+		$this->geocoder = MapsFactory::newDefault()->newGeocoder();
+	}
 
 	/**
 	 * @since 3.1
 	 *
 	 * @param string $value
 	 *
-	 * @return WmsOverlay
+	 * @return ImageOverlay
 	 * @throws ParseException
 	 */
-	protected function stringParse( $value ) {
+	public function parse( $value ) {
 		$parameters = explode( '~', $value );
 		$imageParameters = explode( ':', $parameters[0], 3 );
 
@@ -41,7 +45,13 @@ class ImageOverlayParser extends StringValueParser {
 	}
 
 	private function stringToLatLongValue( $location ) {
-		return ( new LatLongParser() )->parse( $location );
+		$latLong = $this->geocoder->geocode( $location );
+
+		if ( $location === null ) {
+			throw new ParseException( 'Failed to parse or geocode' );
+		}
+
+		return $latLong;
 	}
 
 }