+++ src/Elements/Location.php
@@ -3,8 +3,6 @@
 namespace Maps\Elements;
 
 use DataValues\Geo\Values\LatLongValue;
-use Maps\Geocoders;
-use MWException;
 
 /**
  * Class describing a single location (geographical point).
@@ -61,44 +59,6 @@ class Location extends BaseElement {
 	 */
 	public static function newFromLatLon( $lat, $lon ) {
 		return new self( new LatLongValue( $lat, $lon ) );
-	}
-
-
-	/**
-	 * Creates and returns a new instance of a Location with title from a latitude and longitude.
-	 *
-	 * @since 3.7
-	 *
-	 * @param float $lat
-	 * @param float $lon
-	 *
-	 * @return Location
-	 */
-	public static function newTitledFromLatLon( $lat, $lon ) {
-		$location = new self( new LatLongValue( $lat, $lon ) );
-		$location->setTitle( $lat . ',' . $lon );
-		return $location;
-	}
-
-	/**
-	 * Creates and returns a new instance of a Location from an address.
-	 *
-	 * @since 1.0
-	 *
-	 * @param string $address
-	 * @deprecated
-	 *
-	 * @return Location
-	 * @throws MWException
-	 */
-	public static function newFromAddress( $address ) {
-		$address = Geocoders::attemptToGeocode( $address );
-
-		if ( $address === false ) {
-			throw new MWException( 'Could not geocode address' );
-		}
-
-		return new static( $address );
 	}
 
 	public function __construct( LatLongValue $coordinates ) {